Add validation to Meal CRUD methods.

+++ FoodService/FoodService/Controllers/MealController.cs
@@ -2,6 +2,7 @@ using System.Threading.Tasks;
 using FoodService.Models;
 using FoodService.Models.RequestModels.Restaurant;
 using FoodService.Services.MealService;
+using FoodService.Services.RestaurantService;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -11,16 +12,23 @@ namespace FoodService.Controllers
     public class MealController : Controller
     {
         private readonly IMealService mealService;
+        private readonly IRestaurantService restaurantService;
 
-        public MealController(IMealService mealService)
+        public MealController(IMealService mealService, IRestaurantService restaurantService)
         {
             this.mealService = mealService;
+            this.restaurantService = restaurantService;
         }
 
         [Authorize(Roles = "Manager, Admin")]
         [HttpGet]
-        public IActionResult Add(long id)
+        public async Task<IActionResult> Add(long id)
         {
+            if (!await restaurantService.ValidateAccessAsync(id, User.Identity.Name))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
+
             AddMealRequest addMealRequest = new AddMealRequest()
             {
                 RestaurantId = id,
@@ -33,6 +41,11 @@ namespace FoodService.Controllers
         [HttpPost]
         public async Task<IActionResult> Add(AddMealRequest addMealRequest)
         {
+            if (!await restaurantService.ValidateAccessAsync(addMealRequest.RestaurantId, User.Identity.Name))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
+
             if (ModelState.IsValid)
             {
                 await mealService.SaveMealAsync(addMealRequest);
@@ -45,6 +58,11 @@ namespace FoodService.Controllers
         [HttpPost]
         public async Task<IActionResult> Delete(long id)
         {
+            if (!await mealService.ValidateAccessAsync(id, User.Identity.Name))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
+
             var meal = await mealService.GetMealByIdAsync(id);
             await mealService.DeleteMealAsync(id);
             return RedirectToAction(nameof(RestaurantController.Edit), "Restaurant", new {id = meal.Restaurant.RestaurantId});
@@ -54,7 +72,12 @@ namespace FoodService.Controllers
         [HttpGet]
         public async Task<IActionResult> Edit(long id)
         {
-           var requestModel = await mealService.CreateMealRequestAsync(id);
+            if (!await mealService.ValidateAccessAsync(id, User.Identity.Name))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
+
+            var requestModel = await mealService.CreateMealRequestAsync(id);
             if(requestModel == null)
             {
                 return RedirectToAction(nameof(HomeController.Index), "Home", new { page = 1 });
@@ -66,6 +89,11 @@ namespace FoodService.Controllers
         [HttpPost]
         public async Task<IActionResult> Edit(AddMealRequest addMealRequest, long id)
         {
+            if (!await mealService.ValidateAccessAsync(id, User.Identity.Name))
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
+
             if (ModelState.IsValid)
             {
                 var meal = await mealService.GetMealByIdAsync(id);
